% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banten.R
\docType{data}
\name{Banten}
\alias{Banten}
\title{Food Prices in Banten}
\format{
A data frame with 32 variables:
\describe{
\item{Tanggal}{Date}
\item{Beras}{Price of rice in Banten}
\item{Beras Kualitas Bawah I}{Price of lower quality rice I in Banten}
\item{Beras Kualitas Bawah II}{Price of lower quality rice II in Banten}
\item{Beras Kualitas Medium I}{Price of medium quality rice I in Banten}
\item{Beras Kualitas Medium II}{Price of medium quality rice II in Banten}
\item{Beras Kualitas Super I}{Price of super quality rice I in Banten}
\item{Beras Kualitas Super II}{Price of super quality rice II in Banten}
\item{Daging Ayam}{Price of chicken meat in Banten}
\item{Daging Ayam Ras Segar}{Price of fresh purebred chicken meat in Banten}
\item{Daging Sapi}{Price of beef in Banten}
\item{Daging Sapi Kualitas 1}{Price of 1st quality beef in Banten}
\item{Daging Sapi Kualitas 2}{Price of 2nd quality beef in Banten}
\item{Telur Ayam}{Price of chicken eggs in Banten}
\item{Telur Ayam Ras Segar}{Price of fresh purebred chicken eggs in Banten}
\item{Bawang Merah}{Price of shallots in Banten}
\item{Bawang Merah Ukuran Sedang}{Price of medium shallots in Banten}
\item{Bawang Putih}{Price of garlic in Banten}
\item{Bawang Putih Ukuran Sedang}{Price of medium garlic in Banten}
\item{Cabai Merah}{Price of red chili in Banten}
\item{Cabai Merah Besar}{Price of big red chili in Banten}
\item{Cabai Merah Keriting}{Price of curly red chilies in Banten}
\item{Cabai Rawit}{Price of Tabasco chili pepper in Banten}
\item{Cabai Rawit Hijau}{Price of green Tabasco chili pepper in Banten}
\item{Cabai Rawit Merah}{Price of green Tabasco chili pepper in Banten}
\item{Minyak Goreng}{Price of cooking oil in Banten}
\item{Minyak Goreng Curah}{Price of bulk cooking oil in Banten}
\item{Minyak Goreng Kemasan Bermerk 1}{Price of branded packaged cooking oil 1 in Banten}
\item{Minyak Goreng Kemasan Bermerk 2}{Price of branded packaged cooking oil 2 in Banten}
\item{Gula Pasir}{Price of sugar in Banten}
\item{Gula Pasir Kualitas Premium}{Price of premium sugar in Banten}
\item{Gula Pasir Lokal}{Price of local sugar in Banten}
}
}
\source{
\describe{
https://drive.google.com/drive/folders/1n2LPPd5PKTJXc6laQf0a0CWPCLKLjNYJ
}
}
\usage{
Banten
}
\description{
Food Prices in Banten
}
\examples{
data(Banten)
}
\keyword{datasets}
