% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIYogyakarta.R
\docType{data}
\name{DIYogyakarta}
\alias{DIYogyakarta}
\title{Food Prices in DI Yogyakarta}
\format{
A data frame with 32 variables:
\describe{
\item{Tanggal}{Date}
\item{Beras}{Price of rice in DI Yogyakarta}
\item{Beras Kualitas Bawah I}{Price of lower quality rice I in DI Yogyakarta}
\item{Beras Kualitas Bawah II}{Price of lower quality rice II in DI Yogyakarta}
\item{Beras Kualitas Medium I}{Price of medium quality rice I in DI Yogyakarta}
\item{Beras Kualitas Medium II}{Price of medium quality rice II in DI Yogyakarta}
\item{Beras Kualitas Super I}{Price of super quality rice I in DI Yogyakarta}
\item{Beras Kualitas Super II}{Price of super quality rice II in DI Yogyakarta}
\item{Daging Ayam}{Price of chicken meat in DI Yogyakarta}
\item{Daging Ayam Ras Segar}{Price of fresh purebred chicken meat in DI Yogyakarta}
\item{Daging Sapi}{Price of beef in DI Yogyakarta}
\item{Daging Sapi Kualitas 1}{Price of 1st quality beef in DI Yogyakarta}
\item{Daging Sapi Kualitas 2}{Price of 2nd quality beef in DI Yogyakarta}
\item{Telur Ayam}{Price of chicken eggs in DI Yogyakarta}
\item{Telur Ayam Ras Segar}{Price of fresh purebred chicken eggs in DI Yogyakarta}
\item{Bawang Merah}{Price of shallots in DI Yogyakarta}
\item{Bawang Merah Ukuran Sedang}{Price of medium shallots in DI Yogyakarta}
\item{Bawang Putih}{Price of garlic in DI Yogyakarta}
\item{Bawang Putih Ukuran Sedang}{Price of medium garlic in DI Yogyakarta}
\item{Cabai Merah}{Price of red chili in DI Yogyakarta}
\item{Cabai Merah Besar}{Price of big red chili in DI Yogyakarta}
\item{Cabai Merah Keriting}{Price of curly red chilies in DI Yogyakarta}
\item{Cabai Rawit}{Price of Tabasco chili pepper in DI Yogyakarta}
\item{Cabai Rawit Hijau}{Price of green Tabasco chili pepper in DI Yogyakarta}
\item{Cabai Rawit Merah}{Price of green Tabasco chili pepper in DI Yogyakarta}
\item{Minyak Goreng}{Price of cooking oil in DI Yogyakarta}
\item{Minyak Goreng Curah}{Price of bulk cooking oil in DI Yogyakarta}
\item{Minyak Goreng Kemasan Bermerk 1}{Price of branded packaged cooking oil 1 in DI Yogyakarta}
\item{Minyak Goreng Kemasan Bermerk 2}{Price of branded packaged cooking oil 2 in DI Yogyakarta}
\item{Gula Pasir}{Price of sugar in DI Yogyakarta}
\item{Gula Pasir Kualitas Premium}{Price of premium sugar in DI Yogyakarta}
\item{Gula Pasir Lokal}{Price of local sugar in DI Yogyakarta}
}
}
\source{
\describe{
https://drive.google.com/drive/folders/1n2LPPd5PKTJXc6laQf0a0CWPCLKLjNYJ
}
}
\usage{
DIYogyakarta
}
\description{
Food Prices in DI Yogyakarta
}
\examples{
data(DIYogyakarta)
}
\keyword{datasets}
