% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DKIJakarta.R
\docType{data}
\name{DKIJakarta}
\alias{DKIJakarta}
\title{Food Prices in DKI Jakarta}
\format{
A data frame with 32 variables:
\describe{
\item{Tanggal}{Date}
\item{Beras}{Price of rice in DKI Jakarta}
\item{Beras Kualitas Bawah I}{Price of lower quality rice I in DKI Jakarta}
\item{Beras Kualitas Bawah II}{Price of lower quality rice II in DKI Jakarta}
\item{Beras Kualitas Medium I}{Price of medium quality rice I in DKI Jakarta}
\item{Beras Kualitas Medium II}{Price of medium quality rice II in DKI Jakarta}
\item{Beras Kualitas Super I}{Price of super quality rice I in DKI Jakarta}
\item{Beras Kualitas Super II}{Price of super quality rice II in DKI Jakarta}
\item{Daging Ayam}{Price of chicken meat in DKI Jakarta}
\item{Daging Ayam Ras Segar}{Price of fresh purebred chicken meat in DKI Jakarta}
\item{Daging Sapi}{Price of beef in DKI Jakarta}
\item{Daging Sapi Kualitas 1}{Price of 1st quality beef in DKI Jakarta}
\item{Daging Sapi Kualitas 2}{Price of 2nd quality beef in DKI Jakarta}
\item{Telur Ayam}{Price of chicken eggs in DKI Jakarta}
\item{Telur Ayam Ras Segar}{Price of fresh purebred chicken eggs in DKI Jakarta}
\item{Bawang Merah}{Price of shallots in DKI Jakarta}
\item{Bawang Merah Ukuran Sedang}{Price of medium shallots in DKI Jakarta}
\item{Bawang Putih}{Price of garlic in DKI Jakarta}
\item{Bawang Putih Ukuran Sedang}{Price of medium garlic in DKI Jakarta}
\item{Cabai Merah}{Price of red chili in DKI Jakarta}
\item{Cabai Merah Besar}{Price of big red chili in DKI Jakarta}
\item{Cabai Merah Keriting}{Price of curly red chilies in DKI Jakarta}
\item{Cabai Rawit}{Price of Tabasco chili pepper in DKI Jakarta}
\item{Cabai Rawit Hijau}{Price of green Tabasco chili pepper in DKI Jakarta}
\item{Cabai Rawit Merah}{Price of green Tabasco chili pepper in DKI Jakarta}
\item{Minyak Goreng}{Price of cooking oil in DKI Jakarta}
\item{Minyak Goreng Curah}{Price of bulk cooking oil in DKI Jakarta}
\item{Minyak Goreng Kemasan Bermerk 1}{Price of branded packaged cooking oil 1 in DKI Jakarta}
\item{Minyak Goreng Kemasan Bermerk 2}{Price of branded packaged cooking oil 2 in DKI Jakarta}
\item{Gula Pasir}{Price of sugar in DKI Jakarta}
\item{Gula Pasir Kualitas Premium}{Price of premium sugar in DKI Jakarta}
\item{Gula Pasir Lokal}{Price of local sugar in DKI Jakarta}
}
}
\source{
\describe{
https://drive.google.com/drive/folders/1n2LPPd5PKTJXc6laQf0a0CWPCLKLjNYJ
}
}
\usage{
DKIJakarta
}
\description{
Food Prices in DKI Jakarta
}
\examples{
data(DKIJakarta)
}
\keyword{datasets}
