% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifpd.R
\name{ifpd}
\alias{ifpd}
\title{Creates Indonesia Food Prices Data}
\usage{
ifpd(data)
}
\arguments{
\item{data}{dataframes}
}
\value{
A dataframes of Indonesia Food Prices Data with imputation
}
\description{
Creates Indonesia Food Prices Data
}
\examples{
ifpd(Aceh)
ifpd(Bali)
ifpd(Banten)
ifpd(Bengkulu)
ifpd(DIYogyakarta)
ifpd(DKIJakarta)
ifpd(Gorontalo)
ifpd(Jambi)
ifpd(JawaBarat)
ifpd(JawaTengah)
ifpd(JawaTimur)
ifpd(KalimantanBarat)
ifpd(KalimantanSelatan)
ifpd(KalimantanTengah)
ifpd(KalimantanTimur)
ifpd(KalimantanUtara)
ifpd(KepulauanBangkaBelitung)
ifpd(KepulauanRiau)
ifpd(Lampung)
ifpd(MalukuUtara)
ifpd(Maluku)
ifpd(NusaTenggaraBarat)
ifpd(NusaTenggaraTimur)
ifpd(PapuaBarat)
ifpd(Riau)
ifpd(SulawesiBarat)
ifpd(SulawesiSelatan)
ifpd(SulawesiTengah)
ifpd(SulawesiTenggara)
ifpd(SulawesiUtara)
ifpd(SumateraBarat)
ifpd(SumateraSelatan)
ifpd(SumateraUtara)

}
