% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jambi.R
\docType{data}
\name{Jambi}
\alias{Jambi}
\title{Food Prices in Jambi}
\format{
A data frame with 31 variables:
\describe{
\item{Tanggal}{Date}
\item{Beras}{Price of rice in Jambi}
\item{Beras Kualitas Bawah I}{Price of lower quality rice I in Jambi}
\item{Beras Kualitas Bawah II}{Price of lower quality rice II in Jambi}
\item{Beras Kualitas Medium I}{Price of medium quality rice I in Jambi}
\item{Beras Kualitas Medium II}{Price of medium quality rice II in Jambi}
\item{Beras Kualitas Super I}{Price of super quality rice I in Jambi}
\item{Beras Kualitas Super II}{Price of super quality rice II in Jambi}
\item{Daging Ayam}{Price of chicken meat in Jambi}
\item{Daging Ayam Ras Segar}{Price of fresh purebred chicken meat in Jambi}
\item{Daging Sapi}{Price of beef in Jambi}
\item{Daging Sapi Kualitas 1}{Price of 1st quality beef in Jambi}
\item{Telur Ayam}{Price of chicken eggs in Jambi}
\item{Telur Ayam Ras Segar}{Price of fresh purebred chicken eggs in Jambi}
\item{Bawang Merah}{Price of shallots in Jambi}
\item{Bawang Merah Ukuran Sedang}{Price of medium shallots in Jambi}
\item{Bawang Putih}{Price of garlic in Jambi}
\item{Bawang Putih Ukuran Sedang}{Price of medium garlic in Jambi}
\item{Cabai Merah}{Price of red chili in Jambi}
\item{Cabai Merah Besar}{Price of big red chili in Jambi}
\item{Cabai Merah Keriting}{Price of curly red chilies in Jambi}
\item{Cabai Rawit}{Price of Tabasco chili pepper in Jambi}
\item{Cabai Rawit Hijau}{Price of green Tabasco chili pepper in Jambi}
\item{Cabai Rawit Merah}{Price of green Tabasco chili pepper in Jambi}
\item{Minyak Goreng}{Price of cooking oil in Jambi}
\item{Minyak Goreng Curah}{Price of bulk cooking oil in Jambi}
\item{Minyak Goreng Kemasan Bermerk 1}{Price of branded packaged cooking oil 1 in Jambi}
\item{Minyak Goreng Kemasan Bermerk 2}{Price of branded packaged cooking oil 2 in Jambi}
\item{Gula Pasir}{Price of sugar in Jambi}
\item{Gula Pasir Kualitas Premium}{Price of premium sugar in Jambi}
\item{Gula Pasir Lokal}{Price of local sugar in Jambi}
}
}
\source{
\describe{
https://drive.google.com/drive/folders/1n2LPPd5PKTJXc6laQf0a0CWPCLKLjNYJ
}
}
\usage{
Jambi
}
\description{
Food Prices in Jambi
}
\examples{
data(Jambi)
}
\keyword{datasets}
