% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JawaBarat.R
\docType{data}
\name{JawaBarat}
\alias{JawaBarat}
\title{Food Prices in Jawa Barat}
\format{
A data frame with 32 variables:
\describe{
\item{Tanggal}{Date}
\item{Beras}{Price of rice in Jawa Barat}
\item{Beras Kualitas Bawah I}{Price of lower quality rice I in Jawa Barat}
\item{Beras Kualitas Bawah II}{Price of lower quality rice II in Jawa Barat}
\item{Beras Kualitas Medium I}{Price of medium quality rice I in Jawa Barat}
\item{Beras Kualitas Medium II}{Price of medium quality rice II in Jawa Barat}
\item{Beras Kualitas Super I}{Price of super quality rice I in Jawa Barat}
\item{Beras Kualitas Super II}{Price of super quality rice II in Jawa Barat}
\item{Daging Ayam}{Price of chicken meat in Jawa Barat}
\item{Daging Ayam Ras Segar}{Price of fresh purebred chicken meat in Jawa Barat}
\item{Daging Sapi}{Price of beef in Jawa Barat}
\item{Daging Sapi Kualitas 1}{Price of 1st quality beef in Jawa Barat}
\item{Daging Sapi Kualitas 2}{Price of 2nd quality beef in Jawa Barat}
\item{Telur Ayam}{Price of chicken eggs in Jawa Barat}
\item{Telur Ayam Ras Segar}{Price of fresh purebred chicken eggs in Jawa Barat}
\item{Bawang Merah}{Price of shallots in Jawa Barat}
\item{Bawang Merah Ukuran Sedang}{Price of medium shallots in Jawa Barat}
\item{Bawang Putih}{Price of garlic in Jawa Barat}
\item{Bawang Putih Ukuran Sedang}{Price of medium garlic in Jawa Barat}
\item{Cabai Merah}{Price of red chili in Jawa Barat}
\item{Cabai Merah Besar}{Price of big red chili in Jawa Barat}
\item{Cabai Merah Keriting}{Price of curly red chilies in Jawa Barat}
\item{Cabai Rawit}{Price of Tabasco chili pepper in Jawa Barat}
\item{Cabai Rawit Hijau}{Price of green Tabasco chili pepper in Jawa Barat}
\item{Cabai Rawit Merah}{Price of green Tabasco chili pepper in Jawa Barat}
\item{Minyak Goreng}{Price of cooking oil in Jawa Barat}
\item{Minyak Goreng Curah}{Price of bulk cooking oil in Jawa Barat}
\item{Minyak Goreng Kemasan Bermerk 1}{Price of branded packaged cooking oil 1 in Jawa Barat}
\item{Minyak Goreng Kemasan Bermerk 2}{Price of branded packaged cooking oil 2 in Jawa Barat}
\item{Gula Pasir}{Price of sugar in Jawa Barat}
\item{Gula Pasir Kualitas Premium}{Price of premium sugar in Jawa Barat}
\item{Gula Pasir Lokal}{Price of local sugar in Jawa Barat}
}
}
\source{
\describe{
https://drive.google.com/drive/folders/1n2LPPd5PKTJXc6laQf0a0CWPCLKLjNYJ
}
}
\usage{
JawaBarat
}
\description{
Food Prices in Jawa Barat
}
\examples{
data(JawaBarat)
}
\keyword{datasets}
