% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KalimantanBarat.R
\docType{data}
\name{KalimantanBarat}
\alias{KalimantanBarat}
\title{Food Prices in Kalimantan Barat}
\format{
A data frame with 30 variables:
\describe{
\item{Tanggal}{Date}
\item{Beras}{Price of rice in Kalimantan Barat}
\item{Beras Kualitas Medium I}{Price of medium quality rice I in Kalimantan Barat}
\item{Beras Kualitas Medium II}{Price of medium quality rice II in Kalimantan Barat}
\item{Beras Kualitas Super I}{Price of super quality rice I in Kalimantan Barat}
\item{Beras Kualitas Super II}{Price of super quality rice II in Kalimantan Barat}
\item{Daging Ayam}{Price of chicken meat in Kalimantan Barat}
\item{Daging Ayam Ras Segar}{Price of fresh purebred chicken meat in Kalimantan Barat}
\item{Daging Sapi}{Price of beef in Kalimantan Barat}
\item{Daging Sapi Kualitas 1}{Price of 1st quality beef in Kalimantan Barat}
\item{Daging Sapi Kualitas 2}{Price of 2nd quality beef in Kalimantan Barat}
\item{Telur Ayam}{Price of chicken eggs in Kalimantan Barat}
\item{Telur Ayam Ras Segar}{Price of fresh purebred chicken eggs in Kalimantan Barat}
\item{Bawang Merah}{Price of shallots in Kalimantan Barat}
\item{Bawang Merah Ukuran Sedang}{Price of medium shallots in Kalimantan Barat}
\item{Bawang Putih}{Price of garlic in Kalimantan Barat}
\item{Bawang Putih Ukuran Sedang}{Price of medium garlic in Kalimantan Barat}
\item{Cabai Merah}{Price of red chili in Kalimantan Barat}
\item{Cabai Merah Besar}{Price of big red chili in Kalimantan Barat}
\item{Cabai Merah Keriting}{Price of curly red chilies in Kalimantan Barat}
\item{Cabai Rawit}{Price of Tabasco chili pepper in Kalimantan Barat}
\item{Cabai Rawit Hijau}{Price of green Tabasco chili pepper in Kalimantan Barat}
\item{Cabai Rawit Merah}{Price of green Tabasco chili pepper in Kalimantan Barat}
\item{Minyak Goreng}{Price of cooking oil in Kalimantan Barat}
\item{Minyak Goreng Curah}{Price of bulk cooking oil in Kalimantan Barat}
\item{Minyak Goreng Kemasan Bermerk 1}{Price of branded packaged cooking oil 1 in Kalimantan Barat}
\item{Minyak Goreng Kemasan Bermerk 2}{Price of branded packaged cooking oil 2 in Kalimantan Barat}
\item{Gula Pasir}{Price of sugar in Kalimantan Barat}
\item{Gula Pasir Kualitas Premium}{Price of premium sugar in Kalimantan Barat}
\item{Gula Pasir Lokal}{Price of local sugar in Kalimantan Barat}
}
}
\source{
\describe{
https://drive.google.com/drive/folders/1n2LPPd5PKTJXc6laQf0a0CWPCLKLjNYJ
}
}
\usage{
KalimantanBarat
}
\description{
Food Prices in Kalimantan Barat
}
\examples{
data(KalimantanBarat)
}
\keyword{datasets}
