% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KalimantanSelatan.R
\docType{data}
\name{KalimantanSelatan}
\alias{KalimantanSelatan}
\title{Food Prices in Kalimantan Selatan}
\format{
A data frame with 32 variables:
\describe{
\item{Tanggal}{Date}
\item{Beras}{Price of rice in Kalimantan Selatan}
\item{Beras Kualitas Bawah I}{Price of lower quality rice I in Kalimantan Selatan}
\item{Beras Kualitas Bawah II}{Price of lower quality rice II in Kalimantan Selatan}
\item{Beras Kualitas Medium I}{Price of medium quality rice I in Kalimantan Selatan}
\item{Beras Kualitas Medium II}{Price of medium quality rice II in Kalimantan Selatan}
\item{Beras Kualitas Super I}{Price of super quality rice I in Kalimantan Selatan}
\item{Beras Kualitas Super II}{Price of super quality rice II in Kalimantan Selatan}
\item{Daging Ayam}{Price of chicken meat in Kalimantan Selatan}
\item{Daging Ayam Ras Segar}{Price of fresh purebred chicken meat in Kalimantan Selatan}
\item{Daging Sapi}{Price of beef in Kalimantan Selatan}
\item{Daging Sapi Kualitas 1}{Price of 1st quality beef in Kalimantan Selatan}
\item{Daging Sapi Kualitas 2}{Price of 2nd quality beef in Kalimantan Selatan}
\item{Telur Ayam}{Price of chicken eggs in Kalimantan Selatan}
\item{Telur Ayam Ras Segar}{Price of fresh purebred chicken eggs in Kalimantan Selatan}
\item{Bawang Merah}{Price of shallots in Kalimantan Selatan}
\item{Bawang Merah Ukuran Sedang}{Price of medium shallots in Kalimantan Selatan}
\item{Bawang Putih}{Price of garlic in Kalimantan Selatan}
\item{Bawang Putih Ukuran Sedang}{Price of medium garlic in Kalimantan Selatan}
\item{Cabai Merah}{Price of red chili in Kalimantan Selatan}
\item{Cabai Merah Besar}{Price of big red chili in Kalimantan Selatan}
\item{Cabai Merah Keriting}{Price of curly red chilies in Kalimantan Selatan}
\item{Cabai Rawit}{Price of Tabasco chili pepper in Kalimantan Selatan}
\item{Cabai Rawit Hijau}{Price of green Tabasco chili pepper in Kalimantan Selatan}
\item{Cabai Rawit Merah}{Price of green Tabasco chili pepper in Kalimantan Selatan}
\item{Minyak Goreng}{Price of cooking oil in Kalimantan Selatan}
\item{Minyak Goreng Curah}{Price of bulk cooking oil in Kalimantan Selatan}
\item{Minyak Goreng Kemasan Bermerk 1}{Price of branded packaged cooking oil 1 in Kalimantan Selatan}
\item{Minyak Goreng Kemasan Bermerk 2}{Price of branded packaged cooking oil 2 in Kalimantan Selatan}
\item{Gula Pasir}{Price of sugar in Kalimantan Selatan}
\item{Gula Pasir Kualitas Premium}{Price of premium sugar in Kalimantan Selatan}
\item{Gula Pasir Lokal}{Price of local sugar in Kalimantan Selatan}
}
}
\source{
\describe{
https://drive.google.com/drive/folders/1n2LPPd5PKTJXc6laQf0a0CWPCLKLjNYJ
}
}
\usage{
KalimantanSelatan
}
\description{
Food Prices in Kalimantan Selatan
}
\examples{
data(KalimantanSelatan)
}
\keyword{datasets}
