% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KepulauanBangkaBelitung.R
\docType{data}
\name{KepulauanBangkaBelitung}
\alias{KepulauanBangkaBelitung}
\title{Food Prices in Kepulauan Bangka Belitung}
\format{
A data frame with 31 variables:
\describe{
\item{Tanggal}{Date}
\item{Beras}{Price of rice in Kepulauan Bangka Belitung}
\item{Beras Kualitas Bawah I}{Price of lower quality rice I in Kepulauan Bangka Belitung}
\item{Beras Kualitas Bawah II}{Price of lower quality rice II in Kepulauan Bangka Belitung}
\item{Beras Kualitas Medium I}{Price of medium quality rice I in Kepulauan Bangka Belitung}
\item{Beras Kualitas Medium II}{Price of medium quality rice II in Kepulauan Bangka Belitung}
\item{Beras Kualitas Super I}{Price of super quality rice I in Kepulauan Bangka Belitung}
\item{Beras Kualitas Super II}{Price of super quality rice II in Kepulauan Bangka Belitung}
\item{Daging Ayam}{Price of chicken meat in Kepulauan Bangka Belitung}
\item{Daging Ayam Ras Segar}{Price of fresh purebred chicken meat in Kepulauan Bangka Belitung}
\item{Daging Sapi}{Price of beef in Kepulauan Bangka Belitung}
\item{Daging Sapi Kualitas 1}{Price of 1st quality beef in Kepulauan Bangka Belitung}
\item{Daging Sapi Kualitas 2}{Price of 2nd quality beef in Kepulauan Bangka Belitung}
\item{Telur Ayam}{Price of chicken eggs in Kepulauan Bangka Belitung}
\item{Telur Ayam Ras Segar}{Price of fresh purebred chicken eggs in Kepulauan Bangka Belitung}
\item{Bawang Merah}{Price of shallots in Kepulauan Bangka Belitung}
\item{Bawang Merah Ukuran Sedang}{Price of medium shallots in Kepulauan Bangka Belitung}
\item{Bawang Putih}{Price of garlic in Kepulauan Bangka Belitung}
\item{Bawang Putih Ukuran Sedang}{Price of medium garlic in Kepulauan Bangka Belitung}
\item{Cabai Merah}{Price of red chili in Kepulauan Bangka Belitung}
\item{Cabai Merah Besar}{Price of big red chili in Kepulauan Bangka Belitung}
\item{Cabai Merah Keriting}{Price of curly red chilies in Kepulauan Bangka Belitung}
\item{Cabai Rawit}{Price of Tabasco chili pepper in Kepulauan Bangka Belitung}
\item{Cabai Rawit Hijau}{Price of green Tabasco chili pepper in Kepulauan Bangka Belitung}
\item{Cabai Rawit Merah}{Price of green Tabasco chili pepper in Kepulauan Bangka Belitung}
\item{Minyak Goreng}{Price of cooking oil in Kepulauan Bangka Belitung}
\item{Minyak Goreng Kemasan Bermerk 1}{Price of branded packaged cooking oil 1 in Kepulauan Bangka Belitung}
\item{Minyak Goreng Kemasan Bermerk 2}{Price of branded packaged cooking oil 2 in Kepulauan Bangka Belitung}
\item{Gula Pasir}{Price of sugar in Kepulauan Bangka Belitung}
\item{Gula Pasir Kualitas Premium}{Price of premium sugar in Kepulauan Bangka Belitung}
\item{Gula Pasir Lokal}{Price of local sugar in Kepulauan Bangka Belitung}
}
}
\source{
\describe{
https://drive.google.com/drive/folders/1n2LPPd5PKTJXc6laQf0a0CWPCLKLjNYJ
}
}
\usage{
KepulauanBangkaBelitung
}
\description{
Food Prices in Kepulauan Bangka Belitung
}
\examples{
data(KepulauanBangkaBelitung)
}
\keyword{datasets}
