% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KepulauanRiau.R
\docType{data}
\name{KepulauanRiau}
\alias{KepulauanRiau}
\title{Food Prices in Kepulauan Riau}
\format{
A data frame with 28 variables:
\describe{
\item{Tanggal}{Date}
\item{Beras}{Price of rice in Kepulauan Riau}
\item{Beras Kualitas Bawah I}{Price of lower quality rice I in Kepulauan Riau}
\item{Beras Kualitas Bawah II}{Price of lower quality rice II in Kepulauan Riau}
\item{Beras Kualitas Medium I}{Price of medium quality rice I in Kepulauan Riau}
\item{Beras Kualitas Medium II}{Price of medium quality rice II in Kepulauan Riau}
\item{Beras Kualitas Super I}{Price of super quality rice I in Kepulauan Riau}
\item{Beras Kualitas Super II}{Price of super quality rice II in Kepulauan Riau}
\item{Daging Ayam}{Price of chicken meat in Kepulauan Riau}
\item{Daging Ayam Ras Segar}{Price of fresh purebred chicken meat in Kepulauan Riau}
\item{Daging Sapi}{Price of beef in Kepulauan Riau}
\item{Daging Sapi Kualitas 1}{Price of 1st quality beef in Kepulauan Riau}
\item{Telur Ayam}{Price of chicken eggs in Kepulauan Riau}
\item{Telur Ayam Ras Segar}{Price of fresh purebred chicken eggs in Kepulauan Riau}
\item{Bawang Merah}{Price of shallots in Kepulauan Riau}
\item{Bawang Merah Ukuran Sedang}{Price of medium shallots in Kepulauan Riau}
\item{Bawang Putih}{Price of garlic in Kepulauan Riau}
\item{Bawang Putih Ukuran Sedang}{Price of medium garlic in Kepulauan Riau}
\item{Cabai Merah}{Price of red chili in Kepulauan Riau}
\item{Cabai Merah Keriting}{Price of curly red chilies in Kepulauan Riau}
\item{Cabai Rawit}{Price of Tabasco chili pepper in Kepulauan Riau}
\item{Cabai Rawit Hijau}{Price of green Tabasco chili pepper in Kepulauan Riau}
\item{Minyak Goreng}{Price of cooking oil in Kepulauan Riau}
\item{Minyak Goreng Curah}{Price of bulk cooking oil in Kepulauan Riau}
\item{Minyak Goreng Kemasan Bermerk 1}{Price of branded packaged cooking oil 1 in Kepulauan Riau}
\item{Minyak Goreng Kemasan Bermerk 2}{Price of branded packaged cooking oil 2 in Kepulauan Riau}
\item{Gula Pasir}{Price of sugar in Kepulauan Riau}
\item{Gula Pasir Lokal}{Price of local sugar in Kepulauan Riau}
}
}
\source{
\describe{
https://drive.google.com/drive/folders/1n2LPPd5PKTJXc6laQf0a0CWPCLKLjNYJ
}
}
\usage{
KepulauanRiau
}
\description{
Food Prices in Kepulauan Riau
}
\examples{
data(KepulauanRiau)
}
\keyword{datasets}
