% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lampung.R
\docType{data}
\name{Lampung}
\alias{Lampung}
\title{Food Prices in Lampung}
\format{
A data frame with 32 variables:
\describe{
\item{Tanggal}{Date}
\item{Beras}{Price of rice in Lampung}
\item{Beras Kualitas Bawah I}{Price of lower quality rice I in Lampung}
\item{Beras Kualitas Bawah II}{Price of lower quality rice II in Lampung}
\item{Beras Kualitas Medium I}{Price of medium quality rice I in Lampung}
\item{Beras Kualitas Medium II}{Price of medium quality rice II in Lampung}
\item{Beras Kualitas Super I}{Price of super quality rice I in Lampung}
\item{Beras Kualitas Super II}{Price of super quality rice II in Lampung}
\item{Daging Ayam}{Price of chicken meat in Lampung}
\item{Daging Ayam Ras Segar}{Price of fresh purebred chicken meat in Lampung}
\item{Daging Sapi}{Price of beef in Lampung}
\item{Daging Sapi Kualitas 1}{Price of 1st quality beef in Lampung}
\item{Daging Sapi Kualitas 2}{Price of 2nd quality beef in Lampung}
\item{Telur Ayam}{Price of chicken eggs in Lampung}
\item{Telur Ayam Ras Segar}{Price of fresh purebred chicken eggs in Lampung}
\item{Bawang Merah}{Price of shallots in Lampung}
\item{Bawang Merah Ukuran Sedang}{Price of medium shallots in Lampung}
\item{Bawang Putih}{Price of garlic in Lampung}
\item{Bawang Putih Ukuran Sedang}{Price of medium garlic in Lampung}
\item{Cabai Merah}{Price of red chili in Lampung}
\item{Cabai Merah Besar}{Price of big red chili in Lampung}
\item{Cabai Merah Keriting}{Price of curly red chilies in Lampung}
\item{Cabai Rawit}{Price of Tabasco chili pepper in Lampung}
\item{Cabai Rawit Hijau}{Price of green Tabasco chili pepper in Lampung}
\item{Cabai Rawit Merah}{Price of green Tabasco chili pepper in Lampung}
\item{Minyak Goreng}{Price of cooking oil in Lampung}
\item{Minyak Goreng Curah}{Price of bulk cooking oil in Lampung}
\item{Minyak Goreng Kemasan Bermerk 1}{Price of branded packaged cooking oil 1 in Lampung}
\item{Minyak Goreng Kemasan Bermerk 2}{Price of branded packaged cooking oil 2 in Lampung}
\item{Gula Pasir}{Price of sugar in Lampung}
\item{Gula Pasir Kualitas Premium}{Price of premium sugar in Lampung}
\item{Gula Pasir Lokal}{Price of local sugar in Lampung}
}
}
\source{
\describe{
https://drive.google.com/drive/folders/1n2LPPd5PKTJXc6laQf0a0CWPCLKLjNYJ
}
}
\usage{
Lampung
}
\description{
Food Prices in Lampung
}
\examples{
data(Lampung)
}
\keyword{datasets}
