% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MalukuUtara.R
\docType{data}
\name{MalukuUtara}
\alias{MalukuUtara}
\title{Food Prices in Maluku Utara}
\format{
A data frame with 31 variables:
\describe{
\item{Tanggal}{Date}
\item{Beras}{Price of rice in Maluku Utara}
\item{Beras Kualitas Bawah I}{Price of lower quality rice I in Maluku Utara}
\item{Beras Kualitas Bawah II}{Price of lower quality rice II in Maluku Utara}
\item{Beras Kualitas Medium I}{Price of medium quality rice I in Maluku Utara}
\item{Beras Kualitas Medium II}{Price of medium quality rice II in Maluku Utara}
\item{Beras Kualitas Super I}{Price of super quality rice I in Maluku Utara}
\item{Beras Kualitas Super II}{Price of super quality rice II in Maluku Utara}
\item{Daging Ayam}{Price of chicken meat in Maluku Utara}
\item{Daging Ayam Ras Segar}{Price of fresh purebred chicken meat in Maluku Utara}
\item{Daging Sapi}{Price of beef in Maluku Utara}
\item{Daging Sapi Kualitas 1}{Price of 1st quality beef in Maluku Utara}
\item{Telur Ayam}{Price of chicken eggs in Maluku Utara}
\item{Telur Ayam Ras Segar}{Price of fresh purebred chicken eggs in Maluku Utara}
\item{Bawang Merah}{Price of shallots in Maluku Utara}
\item{Bawang Merah Ukuran Sedang}{Price of medium shallots in Maluku Utara}
\item{Bawang Putih}{Price of garlic in Maluku Utara}
\item{Bawang Putih Ukuran Sedang}{Price of medium garlic in Maluku Utara}
\item{Cabai Merah}{Price of red chili in Maluku Utara}
\item{Cabai Merah Besar}{Price of big red chili in Maluku Utara}
\item{Cabai Merah Keriting}{Price of curly red chilies in Maluku Utara}
\item{Cabai Rawit}{Price of Tabasco chili pepper in Maluku Utara}
\item{Cabai Rawit Hijau}{Price of green Tabasco chili pepper in Maluku Utara}
\item{Cabai Rawit Merah}{Price of green Tabasco chili pepper in Maluku Utara}
\item{Minyak Goreng}{Price of cooking oil in Maluku Utara}
\item{Minyak Goreng Curah}{Price of bulk cooking oil in Maluku Utara}
\item{Minyak Goreng Kemasan Bermerk 1}{Price of branded packaged cooking oil 1 in Maluku Utara}
\item{Minyak Goreng Kemasan Bermerk 2}{Price of branded packaged cooking oil 2 in Maluku Utara}
\item{Gula Pasir}{Price of sugar in Maluku Utara}
\item{Gula Pasir Kualitas Premium}{Price of premium sugar in Maluku Utara}
\item{Gula Pasir Lokal}{Price of local sugar in Maluku Utara}
}
}
\source{
\describe{
https://drive.google.com/drive/folders/1n2LPPd5PKTJXc6laQf0a0CWPCLKLjNYJ
}
}
\usage{
MalukuUtara
}
\description{
Food Prices in Maluku Utara
}
\examples{
data(MalukuUtara)
}
\keyword{datasets}
