% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PapuaBarat.R
\docType{data}
\name{PapuaBarat}
\alias{PapuaBarat}
\title{Food Prices in Papua Barat}
\format{
A data frame with 32 variables:
\describe{
\item{Tanggal}{Date}
\item{Beras}{Price of rice in Papua Barat}
\item{Beras Kualitas Bawah I}{Price of lower quality rice I in Papua Barat}
\item{Beras Kualitas Bawah II}{Price of lower quality rice II in Papua Barat}
\item{Beras Kualitas Medium I}{Price of medium quality rice I in Papua Barat}
\item{Beras Kualitas Medium II}{Price of medium quality rice II in Papua Barat}
\item{Beras Kualitas Super I}{Price of super quality rice I in Papua Barat}
\item{Beras Kualitas Super II}{Price of super quality rice II in Papua Barat}
\item{Daging Ayam}{Price of chicken meat in Papua Barat}
\item{Daging Ayam Ras Segar}{Price of fresh purebred chicken meat in Papua Barat}
\item{Daging Sapi}{Price of beef in Papua Barat}
\item{Daging Sapi Kualitas 1}{Price of 1st quality beef in Papua Barat}
\item{Daging Sapi Kualitas 2}{Price of 2nd quality beef in Papua Barat}
\item{Telur Ayam}{Price of chicken eggs in Papua Barat}
\item{Telur Ayam Ras Segar}{Price of fresh purebred chicken eggs in Papua Barat}
\item{Bawang Merah}{Price of shallots in Papua Barat}
\item{Bawang Merah Ukuran Sedang}{Price of medium shallots in Papua Barat}
\item{Bawang Putih}{Price of garlic in Papua Barat}
\item{Bawang Putih Ukuran Sedang}{Price of medium garlic in Papua Barat}
\item{Cabai Merah}{Price of red chili in Papua Barat}
\item{Cabai Merah Besar}{Price of big red chili in Papua Barat}
\item{Cabai Merah Keriting}{Price of curly red chilies in Papua Barat}
\item{Cabai Rawit}{Price of Tabasco chili pepper in Papua Barat}
\item{Cabai Rawit Hijau}{Price of green Tabasco chili pepper in Papua Barat}
\item{Cabai Rawit Merah}{Price of green Tabasco chili pepper in Papua Barat}
\item{Minyak Goreng}{Price of cooking oil in Papua Barat}
\item{Minyak Goreng Curah}{Price of bulk cooking oil in Papua Barat}
\item{Minyak Goreng Kemasan Bermerk 1}{Price of branded packaged cooking oil 1 in Papua Barat}
\item{Minyak Goreng Kemasan Bermerk 2}{Price of branded packaged cooking oil 2 in Papua Barat}
\item{Gula Pasir}{Price of sugar in Papua Barat}
\item{Gula Pasir Kualitas Premium}{Price of premium sugar in Papua Barat}
\item{Gula Pasir Lokal}{Price of local sugar in Papua Barat}
}
}
\source{
\describe{
https://drive.google.com/drive/folders/1n2LPPd5PKTJXc6laQf0a0CWPCLKLjNYJ
}
}
\usage{
PapuaBarat
}
\description{
Food Prices in Papua Barat
}
\examples{
data(PapuaBarat)
}
\keyword{datasets}
