% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SulawesiBarat.R
\docType{data}
\name{SulawesiBarat}
\alias{SulawesiBarat}
\title{Food Prices in Sulawesi Barat}
\format{
A data frame with 32 variables:
\describe{
\item{Tanggal}{Date}
\item{Beras}{Price of rice in Sulawesi Barat}
\item{Beras Kualitas Bawah I}{Price of lower quality rice I in Sulawesi Barat}
\item{Beras Kualitas Bawah II}{Price of lower quality rice II in Sulawesi Barat}
\item{Beras Kualitas Medium I}{Price of medium quality rice I in Sulawesi Barat}
\item{Beras Kualitas Medium II}{Price of medium quality rice II in Sulawesi Barat}
\item{Beras Kualitas Super I}{Price of super quality rice I in Sulawesi Barat}
\item{Beras Kualitas Super II}{Price of super quality rice II in Sulawesi Barat}
\item{Daging Ayam}{Price of chicken meat in Sulawesi Barat}
\item{Daging Ayam Ras Segar}{Price of fresh purebred chicken meat in Sulawesi Barat}
\item{Daging Sapi}{Price of beef in Sulawesi Barat}
\item{Daging Sapi Kualitas 1}{Price of 1st quality beef in Sulawesi Barat}
\item{Daging Sapi Kualitas 2}{Price of 2nd quality beef in Sulawesi Barat}
\item{Telur Ayam}{Price of chicken eggs in Sulawesi Barat}
\item{Telur Ayam Ras Segar}{Price of fresh purebred chicken eggs in Sulawesi Barat}
\item{Bawang Merah}{Price of shallots in Sulawesi Barat}
\item{Bawang Merah Ukuran Sedang}{Price of medium shallots in Sulawesi Barat}
\item{Bawang Putih}{Price of garlic in Sulawesi Barat}
\item{Bawang Putih Ukuran Sedang}{Price of medium garlic in Sulawesi Barat}
\item{Cabai Merah}{Price of red chili in Sulawesi Barat}
\item{Cabai Merah Besar}{Price of big red chili in Sulawesi Barat}
\item{Cabai Merah Keriting}{Price of curly red chilies in Sulawesi Barat}
\item{Cabai Rawit}{Price of Tabasco chili pepper in Sulawesi Barat}
\item{Cabai Rawit Hijau}{Price of green Tabasco chili pepper in Sulawesi Barat}
\item{Cabai Rawit Merah}{Price of green Tabasco chili pepper in Sulawesi Barat}
\item{Minyak Goreng}{Price of cooking oil in Sulawesi Barat}
\item{Minyak Goreng Curah}{Price of bulk cooking oil in Sulawesi Barat}
\item{Minyak Goreng Kemasan Bermerk 1}{Price of branded packaged cooking oil 1 in Sulawesi Barat}
\item{Minyak Goreng Kemasan Bermerk 2}{Price of branded packaged cooking oil 2 in Sulawesi Barat}
\item{Gula Pasir}{Price of sugar in Sulawesi Barat}
\item{Gula Pasir Kualitas Premium}{Price of premium sugar in Sulawesi Barat}
\item{Gula Pasir Lokal}{Price of local sugar in Sulawesi Barat}
}
}
\source{
\describe{
https://drive.google.com/drive/folders/1n2LPPd5PKTJXc6laQf0a0CWPCLKLjNYJ
}
}
\usage{
SulawesiBarat
}
\description{
Food Prices in Sulawesi Barat
}
\examples{
data(SulawesiBarat)
}
\keyword{datasets}
